Move impl of TokensInput
